pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "real-time-fraud-detection"
version = "0.1.0"
description = "Real-time fraud detection system for financial transactions using Apache Kafka, Spark Streaming, and AWS services"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["fraud-detection", "kafka", "spark", "streaming", "aws", "real-time"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
    "Topic :: Database :: Database Engines/Servers",
]
requires-python = ">=3.9"
dependencies = [
    "pyspark>=3.4.1",
    "delta-spark>=2.4.0",
    "pandas>=2.0.3",
    "numpy>=1.24.3",
    "pyarrow>=12.0.1",
    "kafka-python>=2.0.2",
    "confluent-kafka>=2.1.1",
    "avro-python3>=1.10.2",
    "boto3>=1.28.44",
    "botocore>=1.31.44",
    "aws-lambda-runtime-api>=1.0.0",
    "aws-glue-sessions>=0.42.0",
    "psycopg2-binary>=2.9.7",
    "sqlalchemy>=2.0.19",
    "streamlit>=1.25.0",
    "fastapi>=0.100.1",
    "uvicorn>=0.23.2",
    "redis>=4.6.0",
    "celery>=5.3.1",
    "websockets>=11.0.3",
    "pydantic>=2.1.1",
    "marshmallow>=3.20.1",
    "jsonschema>=4.19.0",
    "structlog>=23.1.0",
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.17.1",
    "statsd>=4.0.1",
    "healthcheck>=1.3.5",
    "cryptography>=41.0.3",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "requests>=2.31.0",
    "aiohttp>=3.8.5",
    "faker>=19.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.1",
    "pytest-html>=3.2.0",
    "locust>=2.15.1",
    "testcontainers>=3.7.1",
    "responses>=0.23.3",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "bandit>=1.7.5",
    "pre-commit>=3.4.0",
    "sphinx>=7.1.2",
    "mkdocs>=1.5.2",
    "mkdocs-material>=9.2.8",
    "jupyter>=1.0.0",
    "ipython>=8.14.0",
    "jupyterlab>=4.0.2",
    "notebook>=7.0.2",
    "ipdb>=0.13.13",
    "debugpy>=1.6.7",
    "types-requests>=2.31.0.10",
    "types-PyYAML>=6.0.12.12",
    "types-python-dateutil>=2.8.19.14",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.1.1",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.1",
    "pytest-html>=3.2.0",
    "locust>=2.15.1",
    "testcontainers>=3.7.1",
    "responses>=0.23.3",
]

docs = [
    "sphinx>=7.1.2",
    "mkdocs>=1.5.2",
    "mkdocs-material>=9.2.8",
]

[project.urls]
Homepage = "https://github.com/yourusername/real-time-fraud-detection"
Documentation = "https://github.com/yourusername/real-time-fraud-detection#readme"
Repository = "https://github.com/yourusername/real-time-fraud-detection.git"
"Bug Tracker" = "https://github.com/yourusername/real-time-fraud-detection/issues"
"Change Log" = "https://github.com/yourusername/real-time-fraud-detection/blob/main/CHANGELOG.md"

[project.scripts]
fraud-detection = "src.main:main"
transaction-simulator = "src.transaction_simulator.main:main"
spark-job = "src.spark_streaming.fraud_detection_job:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json", "*.sql"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
known_third_party = ["pyspark", "kafka", "boto3", "fastapi", "pytest"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pyspark.*",
    "delta.*",
    "confluent_kafka.*",
    "boto3.*",
    "botocore.*",
    "psycopg2.*",
    "redis.*",
    "celery.*",
    "prometheus_client.*",
    "structlog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", "env"]
skips = ["B101", "B601"]

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.poetry]
name = "real-time-fraud-detection"
version = "0.1.0"
description = "Real-time fraud detection system for financial transactions"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.0.0"


